Move LikedTweet model to the domain package
Entity models live in domain (Tweet, User, Like, Notification) — the
repo and handler now reference domain.LikedTweet instead of a
database-local struct.
