+++ domain/warpnet.go
@@ -118,6 +118,17 @@ type Like struct {
 	UserId  string `json:"user_id"`
 }
 
+// LikedTweet defines model for LikedTweet — one entry of a user's
+// "tweets I liked" index. OwnerUserId is the tweet author's id, stored
+// alongside so clients can fetch the tweet without an extra resolution
+// round-trip.
+type LikedTweet struct {
+	UserId      string    `json:"user_id"`
+	TweetId     string    `json:"tweet_id"`
+	OwnerUserId string    `json:"owner_user_id"`
+	CreatedAt   time.Time `json:"created_at"`
+}
+
 // Owner defines model for Owner.
 type Owner struct {
 	CreatedAt       time.Time `json:"created_at"`
